+++ test/proxies_test.js
@@ -9,6 +9,7 @@ describe('Automerge proxy API', () => {
       Automerge.change(Automerge.init(), doc => {
         assert.strictEqual(doc._type, 'map')
         assert.strictEqual(doc._objectId, ROOT_ID)
+        assert.strictEqual('_objectId' in doc, true)
       })
     })
 
@@ -17,6 +18,7 @@ describe('Automerge proxy API', () => {
         assert(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/.test(doc._actorId))
         assert.notEqual(doc._actorId, ROOT_ID)
         assert.strictEqual(Automerge.init('customActorId')._actorId, 'customActorId')
+        assert.strictEqual('_actorId' in doc, true)
       })
     })
 
@@ -45,57 +47,56 @@ describe('Automerge proxy API', () => {
 
     it('should support Object.keys()', () => {
       Automerge.change(Automerge.init(), doc => {
-        assert.deepEqual(Object.keys(doc), ['_objectId'])
+        assert.deepEqual(Object.keys(doc), [])
         doc.key1 = 'value1'
-        equalsOneOf(Object.keys(doc), ['_objectId', 'key1'], ['key1', '_objectId'])
+        assert.deepEqual(Object.keys(doc), ['key1'])
         doc.key2 = 'value2'
-        equalsOneOf(Object.keys(doc),
-          ['_objectId', 'key1', 'key2'], ['_objectId', 'key2', 'key1'],
-          ['key1', '_objectId', 'key2'], ['key2', '_objectId', 'key1'],
-          ['key1', 'key2', '_objectId'], ['key2', 'key1', '_objectId'])
+        equalsOneOf(Object.keys(doc), ['key1', 'key2'], ['key2', 'key1'])
       })
     })
 
     it('should support Object.getOwnPropertyNames()', () => {
       Automerge.change(Automerge.init(), doc => {
-        assert.deepEqual(Object.getOwnPropertyNames(doc), ['_objectId'])
+        assert.deepEqual(Object.getOwnPropertyNames(doc), [])
         doc.key1 = 'value1'
-        equalsOneOf(Object.getOwnPropertyNames(doc), ['_objectId', 'key1'], ['key1', '_objectId'])
+        assert.deepEqual(Object.getOwnPropertyNames(doc), ['key1'])
         doc.key2 = 'value2'
-        equalsOneOf(Object.getOwnPropertyNames(doc),
-          ['_objectId', 'key1', 'key2'], ['_objectId', 'key2', 'key1'],
-          ['key1', '_objectId', 'key2'], ['key2', '_objectId', 'key1'],
-          ['key1', 'key2', '_objectId'], ['key2', 'key1', '_objectId'])
+        equalsOneOf(Object.getOwnPropertyNames(doc), ['key1', 'key2'], ['key2', 'key1'])
+      })
+    })
+
+    it('should support bulk assignment with Object.assign()', () => {
+      Automerge.change(Automerge.init(), doc => {
+        Object.assign(doc, {key1: 'value1', key2: 'value2'})
+        assert.deepEqual(doc, {key1: 'value1', key2: 'value2'})
       })
     })
 
     it('should support JSON.stringify()', () => {
       Automerge.change(Automerge.init(), doc => {
-        assert.deepEqual(JSON.stringify(doc), '{"_objectId":"00000000-0000-0000-0000-000000000000"}')
+        assert.deepEqual(JSON.stringify(doc), '{}')
         doc.key1 = 'value1'
-        equalsOneOf(JSON.stringify(doc),
-          '{"_objectId":"00000000-0000-0000-0000-000000000000","key1":"value1"}',
-          '{"key1":"value1","_objectId":"00000000-0000-0000-0000-000000000000"}')
+        assert.deepEqual(JSON.stringify(doc), '{"key1":"value1"}')
         doc.key2 = 'value2'
         assert.deepEqual(JSON.parse(JSON.stringify(doc)), {
-          _objectId: ROOT_ID, key1: 'value1', key2: 'value2'
+          key1: 'value1', key2: 'value2'
         })
       })
     })
 
     it('should allow inspection as regular JS objects', () => {
       Automerge.change(Automerge.init(), doc => {
-        assert.deepEqual(doc._inspect, {_objectId: ROOT_ID})
-        assert.deepEqual(Automerge.inspect(doc), {_objectId: ROOT_ID})
+        assert.deepEqual(doc._inspect, {})
+        assert.deepEqual(Automerge.inspect(doc), {})
         doc.key1 = 'value1'
-        assert.deepEqual(doc._inspect, {_objectId: ROOT_ID, key1: 'value1'})
-        assert.deepEqual(Automerge.inspect(doc), {_objectId: ROOT_ID, key1: 'value1'})
+        assert.deepEqual(doc._inspect, {key1: 'value1'})
+        assert.deepEqual(Automerge.inspect(doc), {key1: 'value1'})
         doc.key2 = 'value2'
         assert.deepEqual(doc._inspect, {
-          _objectId: ROOT_ID, key1: 'value1', key2: 'value2'
+          key1: 'value1', key2: 'value2'
         })
         assert.deepEqual(Automerge.inspect(doc), {
-          _objectId: ROOT_ID, key1: 'value1', key2: 'value2'
+          key1: 'value1', key2: 'value2'
         })
       })
     })
@@ -164,7 +165,7 @@ describe('Automerge proxy API', () => {
     it('should support JSON.stringify()', () => {
       Automerge.change(root, doc => {
         assert.deepEqual(JSON.parse(JSON.stringify(doc)), {
-          _objectId: ROOT_ID, list: [1, 2, 3], empty: []
+          list: [1, 2, 3], empty: []
         })
         assert.deepEqual(JSON.stringify(doc.list), '[1,2,3]')
       })
@@ -173,10 +174,10 @@ describe('Automerge proxy API', () => {
     it('should allow inspection as regular JS objects', () => {
       Automerge.change(root, doc => {
         assert.deepEqual(doc._inspect, {
-          _objectId: ROOT_ID, list: [1, 2, 3], empty: []
+          list: [1, 2, 3], empty: []
         })
         assert.deepEqual(Automerge.inspect(doc), {
-          _objectId: ROOT_ID, list: [1, 2, 3], empty: []
+          list: [1, 2, 3], empty: []
         })
       })
     })